requirements.txt
pytest
pytest-check
pytest-mock
pymongo==3.11.0
ruamel.yaml
tldextract
python-Levenshtein
fuzzywuzzy
GitPython